+++ src/Asset/PathVersionStrategy.php
@@ -13,21 +13,22 @@ namespace App\Asset;
 
 use Symfony\Component\Asset\VersionStrategy\VersionStrategyInterface;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\HttpKernel\KernelInterface;
 
 class PathVersionStrategy implements VersionStrategyInterface
 {
     private $manifestPath, $manifestData;
 
-    public function __construct(RequestStack $request, string $themesDir, string $defaultTheme, string $adminThemeDir)
+    public function __construct(RequestStack $request, string $defaultTheme, KernelInterface $kernel)
     {
         if (is_null($request) || is_null($request->getCurrentRequest())) {
             return;
         }
         if (preg_match('`^/admin/`', $request->getCurrentRequest()->getPathInfo())) { // Do not add public themes on admin URLs
-            $this->manifestPath = sprintf('%s/build/manifest.json', $adminThemeDir);
+            $this->manifestPath = sprintf('%s/%s/admin/theme/build/manifest.json', $kernel->getProjectDir(), $this->getPublicDirectory($kernel));
         } else {
             $theme = $request->getCurrentRequest()->attributes->get('_theme', $defaultTheme);
-            $this->manifestPath = sprintf('%s/%s/build/manifest.json', $themesDir, $theme);
+            $this->manifestPath = sprintf('%s/%s/themes/%s/build/manifest.json', $kernel->getProjectDir(), $this->getPublicDirectory($kernel), $theme);
         }
     }
 
@@ -55,5 +56,24 @@ class PathVersionStrategy implements VersionStrategyInterface
         }
 
         return isset($this->manifestData[$path]) ? $this->manifestData[$path] : null;
+    }
+
+    private function getPublicDirectory(KernelInterface $kernel): string
+    {
+        $defaultPublicDir = 'public';
+
+        $composerFilePath = $kernel->getProjectDir() . '/composer.json';
+
+        if (!is_readable($composerFilePath)) {
+            return $defaultPublicDir;
+        }
+
+        $composerConfig = json_decode(file_get_contents($composerFilePath), true);
+
+        if (isset($composerConfig['extra']['public-dir'])) {
+            return $composerConfig['extra']['public-dir'];
+        }
+
+        return $defaultPublicDir;
     }
 }